etl/requirements.txt
pandas==1.4.0
numpy==1.22.0
python-dateutil==2.8.2
pytz==2021.3
psycopg2-binary==2.9.3
SQLAlchemy==1.4.31